requirements.txt
pandas>=1.3.0
numpy>=1.20.0
requests>=2.25.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
streamlit>=1.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
tqdm>=4.60.0
python-dotenv>=0.19.0
concurrent-log-handler>=0.9.20  # For better logging with multiprocessing
memory-profiler>=0.60.0  # For memory usage monitoring
psutil>=5.8.0  # For system resource monitoring
